# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jikan API",
...
        "termsOfService": "https://jikan.moe/terms",
        "contact": {
            "name": "API Support (Discord)",
            "url": "http://discord.jikan.moe"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/jikan-me/jikan-rest/blob/master/LICENSE"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "https://api.jikan.moe/v4",
            "description": "Jikan REST API Beta"
        }
    ],
    "paths": {
...
        "/manga/{id}": {
            "get": {
                "tags": [
                    "manga"
                ],
                "operationId": "getMangaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns pictures related to the entry",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/manga"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/manga/{id}/characters": {
            "get": {
                "tags": [
                    "manga"
                ],
                "operationId": "getMangaCharacters",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns manga characters resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/manga_characters"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/manga/{id}/news": {
            "get": {
                "tags": [
                    "manga"
                ],
                "operationId": "getMangaNews",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of manga news topics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/manga_news"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/manga/{id}/forum": {
            "get": {
                "tags": [
                    "manga"
                ],
                "operationId": "getMangaTopics",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter topics",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "all",
                                "episode",
                                "other"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of manga forum topics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/forum"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/manga/{id}/pictures": {
            "get": {
                "tags": [
                    "manga"
                ],
                "operationId": "getMangaPictures",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of manga pictures",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/manga_pictures"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/manga/{id}/statistics": {
            "get": {
                "tags": [
                    "manga"
                ],
                "operationId": "getMangaStatistics",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns anime statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/manga_statistics"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/manga/{id}/moreinfo": {
            "get": {
                "tags": [
                    "manga"
                ],
                "operationId": "getMangaMoreInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns manga moreinfo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/moreinfo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/manga/{id}/recommendations": {
            "get": {
                "tags": [
                    "manga"
                ],
                "operationId": "getMangaRecommendations",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns manga recommendations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/entry_recommendations"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "manga_news": {
                "description": "Manga News Resource",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/pagination"
                    },
                    {
                        "$ref": "#/components/schemas/news"
                    }
                ]
            },
...
            "manga_characters": {
                "description": "Manga Characters Resource",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "character": {
                                    "$ref": "#/components/schemas/character_meta"
                                },
                                "role": {
                                    "description": "Character's Role",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
            "page": {
                "name": "page",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
...
        }
    },
...
}